+++ tests/conftest.py
@@ -19,20 +19,13 @@ from os import environ
 from pathlib import Path
 from sys import platform
 from time import monotonic, sleep
-from unittest.mock import Mock
 from urllib.request import URLError, urlopen
 
-import pytest
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-import mlrun.k8s_utils
-from mlrun.api.db.sqldb.db import run_time_fmt, SQLDB
-from mlrun.config import config
+from mlrun.api.db.sqldb.db import run_time_fmt
 from mlrun.api.db.sqldb.models import Base
-from mlrun.k8s_utils import get_k8s_helper
-from typing import Callable
-from typing import Generator
 
 here = Path(__file__).absolute().parent
 results = here / 'test_results'
@@ -50,55 +43,7 @@ examples_path = Path(here).parent.joinpath('examples')
 environ['PYTHONPATH'] = root_path
 environ['MLRUN_DBPATH'] = rundb_path
 environ['MLRUN_httpdb__dirpath'] = rundb_path
-
-
-@pytest.fixture
-def k8s_helper_mock(monkeypatch):
-    class K8sHelperMock(Mock):
-        def resolve_namespace(self, namespace=None):
-            return namespace or config.namespace
-
-        def is_running_inside_kubernetes_cluster(self):
-            return False
-
-    monkeypatch.setattr(mlrun.k8s_utils, "K8sHelper", K8sHelperMock)
-
-    # call get_k8s_helper so that the mock instance will get into cache
-    k8s_helper_mock_instance = get_k8s_helper()
-
-    yield k8s_helper_mock_instance
-
-    k8s_helper_mock_instance.reset_mock()
-
-
-session_maker: Callable
-
-
-@pytest.fixture
-def db():
-    global session_maker
-    dsn = "sqlite:///:memory:?check_same_thread=false"
-    db_session = None
-    try:
-        session_maker = init_sqldb(dsn)
-        db_session = session_maker()
-        db = SQLDB(dsn)
-        db.initialize(db_session)
-    finally:
-        if db_session is not None:
-            db_session.close()
-    return db
-
-
-@pytest.fixture()
-def db_session() -> Generator:
-    db_session = None
-    try:
-        db_session = session_maker()
-        yield db_session
-    finally:
-        if db_session is not None:
-            db_session.close()
+pytest_plugins = ['tests.common_fixtures']
 
 
 def check_docker():
